+++ src/sortuse64.c
@@ -775,17 +775,21 @@ SEXP r_ram_integer64_ordertab_asc(
               ret[pos++] = ret[i];
           SET_LENGTH(ret_, pos); /* re-allocates ret_ */
       }else{
+        int n_ret = LENGTH(ret_); // allow bailing if user mis-specified nunique (#168)
+        if (n_ret > 0) {
           j = 0;
           ret[j] = 1;
           pos = index[j]-1;
           for(i=1;i<n;i++){
             if (table[index[i]-1]!=table[pos]){
                 pos = index[i]-1;
+                if (j + 1 >= n_ret) break;
                 ret[++j] = 1;
             }else{
                 ret[j]++;
             }
           }
+        }
       }
       PROTECT(ret_); /* Thanks to Tomas Kalibera */
       R_Busy(0);